requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
streamlit==1.28.0
python-dotenv==1.0.0
pydantic==2.5.0
huggingface-hub==0.16.4
numpy<2.0.0,>=1.20.0
openai==1.3.0
transformers==4.30.0
torch==2.0.0
pydantic-settings==2.0.3
aiofiles==23.2.1
python-magic==0.4.27
requests==2.31.0